+++ src/Http/Controllers/ConfirmedPasswordStatusController.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Laravel\Fortify\Http\Controllers;
+
+use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
+
+class ConfirmedPasswordStatusController extends Controller
+{
+    /**
+     * Get the password confirmation status.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Request $request)
+    {
+        return response()->json([
+            'confirmed' => (time() - $request->session()->get('auth.password_confirmed_at', 0)) < $request->input('seconds', config('auth.password_timeout', 900)),
+        ]);
+    }
+}