+++ src/Http/Controllers/TwoFactorAuthenticatedSessionController.php
@@ -6,9 +6,9 @@ use Illuminate\Contracts\Auth\StatefulGuard;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Laravel\Fortify\Contracts\TwoFactorChallengeViewResponse;
+use Laravel\Fortify\Contracts\TwoFactorLoginResponse;
 use Laravel\Fortify\Http\Requests\TwoFactorLoginRequest;
 use Laravel\Fortify\Http\Responses\FailedTwoFactorLoginResponse;
-use Laravel\Fortify\Http\Responses\TwoFactorLoginResponse;
 
 class TwoFactorAuthenticatedSessionController extends Controller
 {