+++ src/Http/Responses/PasswordConfirmedResponse.php
@@ -4,9 +4,9 @@ namespace Laravel\Fortify\Http\Responses;
 
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
-use Laravel\Fortify\Contracts\RegisterResponse as RegisterResponseContract;
+use Laravel\Fortify\Contracts\PasswordConfirmedResponse as PasswordConfirmedResponseContract;
 
-class RegisterResponse implements RegisterResponseContract
+class PasswordConfirmedResponse implements PasswordConfirmedResponseContract
 {
     /**
      * Create an HTTP response that represents the object.
@@ -18,6 +18,6 @@ class RegisterResponse implements RegisterResponseContract
     {
         return $request->wantsJson()
                     ? new JsonResponse('', 201)
-                    : redirect(config('fortify.home'));
+                    : redirect()->intended(config('fortify.home'));
     }
 }