+++ src/Http/Responses/SimpleViewResponse.php
@@ -2,6 +2,7 @@
 
 namespace Laravel\Fortify\Http\Responses;
 
+use Laravel\Fortify\Contracts\ConfirmPasswordViewResponse;
 use Laravel\Fortify\Contracts\LoginViewResponse;
 use Laravel\Fortify\Contracts\RegisterViewResponse;
 use Laravel\Fortify\Contracts\RequestPasswordResetLinkViewResponse;
@@ -15,19 +16,20 @@ class SimpleViewResponse implements
     RegisterViewResponse,
     RequestPasswordResetLinkViewResponse,
     TwoFactorChallengeViewResponse,
-    VerifyEmailViewResponse
+    VerifyEmailViewResponse,
+    ConfirmPasswordViewResponse
 {
     /**
-     * The name of the view.
+     * The name of the view or the callable used to generate the view.
      *
-     * @var string
+     * @var callable|string
      */
     protected $view;
 
     /**
      * Create a new response instance.
      *
-     * @param  string  $view
+     * @param  callable|string  $view
      * @return void
      */
     public function __construct($view)
@@ -43,6 +45,8 @@ class SimpleViewResponse implements
      */
     public function toResponse($request)
     {
-        return view($this->view, ['request' => $request]);
+        return is_callable($this->view) && ! is_string($this->view)
+                    ? call_user_func($this->view, $request)
+                    : view($this->view, ['request' => $request]);
     }
 }